Added GSM info test & moved get_encoding() to types.cpp
